requirements.txt
# Web framework
Flask>=3.1.0

# SQLAlchemy + PostgreSQL
Flask-SQLAlchemy>=3.1.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9

# Cache
Flask-Caching>=2.3.0

# Utilidades
python-dotenv>=1.0.0
requests>=2.32.0
pandas>=2.2.0
MarkupSafe>=2.1.0

# Producción
gunicorn>=23.0.0
